Refactor: Cycles: Modify image loader API
* Move associate_alpha into image metadata
* Change load pixel arguments
* Switch from size_t to int64_t

+++ intern/cycles/util/image.h
@@ -20,14 +20,14 @@ using OIIO::ImageSpec;
 
 template<typename T>
 void util_image_resize_pixels(const vector<T> &input_pixels,
-                              const size_t input_width,
-                              const size_t input_height,
-                              const size_t input_depth,
-                              const size_t components,
+                              const int64_t input_width,
+                              const int64_t input_height,
+                              const int64_t input_depth,
+                              const int64_t components,
                               vector<T> *output_pixels,
-                              size_t *output_width,
-                              size_t *output_height,
-                              size_t *output_depth);
+                              int64_t *output_width,
+                              int64_t *output_height,
+                              int64_t *output_depth);
 
 /* Cast input pixel from unknown storage to float. */
 template<typename T> inline float util_image_cast_to_float(T value);
